Production Version 1 (Malelane)

+++ Traceability/globals.py
@@ -399,18 +399,18 @@ def post_HandlingUnits():
             hu_json = json.load(f)
             f.close()
             print(hu_json)
-            if hu_json['ScannedCode'] != '':
-                hu_json['Status'] = 'From CPC'
+            if hu_json['ScannedCode'] != '':               
                 #res = requests.post(MachineConfiguration.BASE_URL + '/Handling_Units/PAS', json=hu_json, verify=True)
                 try:
-                    res = requests.post(MachineConfiguration.BASE_URL + '/Handling_Units/PAS', json=hu_json, verify=True)
+                    hu_json['Status'] = 'From CPC'
+                    res = requests.post(MachineConfiguration.BASE_URL + '/Handling_Units/PAS', json=hu_json, verify=True)  
+                    json_object = json.dumps(hu_json, indent = 4)
+                    f = open(path + hu, "w")
+                    f.write(json_object)  
+                    f.close()   
                     print('Handling Unit Sync Completed')
                 except:
                     print('Handling Unit Sync Failed')
-                else:
-                    res = requests.post(MachineConfiguration.BASE_URL + '/Handling_Units/PAS', json=hu_json, verify=True)
-                    print('Handling Unit Sync Completed')
-                    #os.remove(path + hu)
                 print(res.text)
             else:
                 os.remove(path + hu)
@@ -450,7 +450,11 @@ def clear_HU():
     path = 'storage/handlingUnits/local/'
     hu_list = os.listdir(path)
     for hu in hu_list:
-        os.remove(path + hu)
+        f = open(path + hu)
+        hu_json = json.load(f)
+        f.close()
+        if hu_json['Status'] == 'From CPC':
+            os.remove(path + hu)
 
 def hu_Reject():
     print('Rejecting HU')
